Org Scope - Initial conversion of provider UI urls to be scoped by organization.
Currently the UI does not allow one user to easily hand another user
a link to an entity. The organization that the entity or a search of
entities is stored in the session and not transmitted through the
URL itself. This change is the first attempt at scoping the provider
URLs based on organization and setting the current organization
based on what exists in the URL instead of the session. This change
does not require all url_for or link_to's to change.