Tests: Refactor pulp glue tests to be more independent of one another.
Previously the glue layer tests were all connected through the use
of support functions with their own cassettes and large collections
of tests tied to single cassettes. This introduces the use of a single
cassette per suite of tests and the removal of dedicated support cassettes.
Each suite of tests, a suite being defined by a TestCase class (not a file),
is reponsible for it's own HTTP requests no matter what entity generates them.
This should reduce the amount of cassette re-generation as updates to Pulp
are made or functionality is added to specific entities.

Previously the feed URL used in tests was dependent upon the system
where the tests were originally ran and disallowed body matching for
repository creation. The feed URL has now been set to '/var/www' and
functionality added to copy the test_repo to this location on every
test run in :all mode. This will allow changes to be made to the repo
and for other developers running the tests to automatically get those
changes. Further, users will no longer need to open up their home
directory to Apache in order to run the tests against a live Pulp.