Fixes #6126: Moves RHSM routes to their own API (/rhsm) and routes file.
Prior, the RHSM routes were anchored at the Katello API and thus
subject to versioning and changes made to the Katello API. However,
the RHSM routes serve as a special set of routes that are outside
the typical API rules and do not version similarly to the Katello ones.
Thus, the RHSM routes have been moved out to their own route file and api
anchor (/rhsm) to allow them to stand alone.